Dictionary Array Functions
turned the pseudo code within the dictionary array functions into actual code. The functions should now be writing nulls and data values to the correct corresponding sensor array. If the array does not exist, it is created. 

The next stage is to actually interface the changes with an apex charts javascript file