Fixing X Axis + comments
The x axis has been updated and fixed to only display numbers as strings instead of floats. There is still the issue that the first value on the x axis has a y value of 0. Everything else is working fine.

Comments were added throughout the code. Additional label options were added but look like garbage with animations turned on.